moves files out of folder
