+++ src/components/games/shark-chase/strategy.js
